Rename client/api/categories.js to collections.js
